Setup basic build glue
Even though we aren't using helix, we are going to need several things
that MRI only exposes as C macros, which helix_runtime sets up for us.
Additionally, `thermite` does not seem to work at all on mac, leaving us
with this as our only existing option for a build system.