Bare minimum implementation of `AttributeSet` and `AttributeSet::Builder`
We actually had to implement quite a bit of `Builder` even to have the
first test passing. It's worth noting that the way `AttributeSet#[]` is
implemented, it is required for memory safety that the value not be held
for longer than the `AttributeSet` itself. Since this object is internal
to Active Record, this is fine. Ironically, Ruby will actually still
track the returned value in its GC, we just don't give it a destructor
function.