Ensure `Attribute.null` still works, implement `Set#[]=`
Right now `_default_attributes` in AR is implemented as
`AttributeSet.new({})`. I don't want to clobber that constant with our
class at the moment, which means that the implementation of the Ruby
`AttributeSet` still needs to work.

Additionally, most failures I see in the AR suite are about `[]=` being
missing.