Allow attributes in the set to out live the set
Before this commit, it was unsafe in Ruby land to save a reference
to a attribute in the set. Since as_ruby() points to the attribute within
the map in Rust land and the set gets dropped when no one refernces the
Ruby handle to the set anymore, Ruby could use freed memory by keeping
a handle from any method that use as_ruby().

The fix is to keep a reference to the set's handle in attributes it
lends out. This extends the life time of the attribute set, which
extends the life time of the attribute enums within the set.
This implementation puts a handle to the parent set in an ivar on
children attributes.

Alternatively, we could allocate all the attribute Ruby handles up
front and keep track of them inside the IndexMap. I went with this
approach because it requires less code change.