Remove as_ruby() usage in attribute's glue
as_ruby() is really unsafe and can easily lead to use-after-free.
The previous implmentation of these methods assume that the lended out
handle built from as_ruby() does not outlive the parent. While this
assumption holds in the intended usage of these methods (Marshal
and Psych), it's better to stay on the safe side.

We will see if it's too slow.