Keep return values from yield alive while mapping
Preivously, it was possible for the GC to free values that were
returned in earlier iterations during mapping. The new
attribute set is not marked by the GC since there isn't a Ruby handle
for it until constrcution of the Rust struct finishes.

The fix is to have a Ruby array on the call stack which keeps all
the values returned from rb_yield. GC scans the call stack and
knows how to traverse a Ruby array.